# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "United Tractor API",
        "description": "United Tractor API documentation",
        "contact": {
            "email": "fajarshidik78@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "View all product",
                "operationId": "615eacadc112c8399b3c3c28eeb93588",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "category_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "kursi"
                                                    },
                                                    "price": {
                                                        "type": "integer",
                                                        "example": 10000
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "IMG-0007"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create a new product",
                "operationId": "9934bcf87d6613c14890dbdd587d311a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "kursi"
                                    },
                                    "price": {
                                        "type": "string",
                                        "example": "10000"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "IMG-0007"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "category_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "kursi"
                                                },
                                                "price": {
                                                    "type": "integer",
                                                    "example": 10000
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "IMG-0007"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "View product by id",
                "operationId": "0f2900456e348f53b24eb29a42c884da",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "category_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "kursi"
                                                },
                                                "price": {
                                                    "type": "integer",
                                                    "example": 10000
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "IMG-0007"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Update a product",
                "operationId": "b4b789a4a3d0c47897f43991a5f12372",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "meja"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 10000
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "IMG-0008"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "category_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "kursi"
                                                },
                                                "price": {
                                                    "type": "integer",
                                                    "example": 10000
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "IMG-0007"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete a product",
                "operationId": "f6c90648962771c8aec7c8aae3d162fa",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "BearerToken": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer &lt;token&gt;)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}